+++ apps/jaspr_pad/lib/providers/project_provider.dart
@@ -67,7 +67,7 @@ final loadedProjectProvider = Provider<AsyncValue<ProjectDataBase>>((ref) {
   }
 
   var storedProject = ref.read(storedProjectProvider);
-  var uri = ComponentsBinding.instance!.currentUri;
+  var uri = ref.binding.currentUri;
   var queryId = uri.queryParameters['gist'] != null
       ? 'gist-${uri.queryParameters['gist']}'
       : uri.queryParameters['sample'] != null
@@ -98,9 +98,11 @@ final loadedProjectProvider = Provider<AsyncValue<ProjectDataBase>>((ref) {
   }
 });
 
-final projectNameProvider = Provider<String?>((ref) => ref.watch(loadedProjectProvider).value?.description);
+final projectNameProvider =
+    Provider<String?>((ref) => ref.watch(loadedProjectProvider).value?.description);
 
-final isTutorialProvider = Provider((ref) => ref.watch(loadedProjectProvider.select((l) => l.value is TutorialData)));
+final isTutorialProvider =
+    Provider((ref) => ref.watch(loadedProjectProvider.select((l) => l.value is TutorialData)));
 
 ProjectDataBase createDefaultProject() {
   return ProjectData(