+++ packages/jaspr/lib/src/server/document/document.dart
@@ -97,7 +97,7 @@ class _FileDocumentState extends State<_FileDocument> {
   @override
   void initState() {
     super.initState();
-    DocumentBinding.instance?._loadFile(component.name);
+    (context.binding as DocumentBinding)._loadFile(component.name);
   }
 
   @override
@@ -129,7 +129,8 @@ class _AppDocumentState extends _ManualDocumentState with _ComponentEntryStateMi
     var id = getIdFor(element);
     _appData = {
       'id': id,
-      if (entry.params != null) 'params': kDebugMode ? entry.params : stateCodec.encode(entry.params),
+      if (entry.params != null)
+        'params': kDebugMode ? entry.params : stateCodec.encode(entry.params),
     };
   }
 
@@ -142,7 +143,11 @@ class _AppDocumentState extends _ManualDocumentState with _ComponentEntryStateMi
 
   @override
   String? get scriptName {
-    var apps = DocumentBinding.instance!._registryElements.entries.where((c) => c.value.isApp).toList();
+    var apps = (context.binding as DocumentBinding)
+        ._registryElements
+        .entries
+        .where((c) => c.value.isApp)
+        .toList();
 
     if (apps.isEmpty) {
       print("[WARNING] Used Document.app() but no app component was provided.");
@@ -172,7 +177,8 @@ class _IslandsDocument extends _ManualDocument {
   State<Document> createState() => _IslandsDocumentState();
 }
 
-class _IslandsDocumentState extends _ManualDocumentState with _ComponentEntryStateMixin<Map<String, String>> {
+class _IslandsDocumentState extends _ManualDocumentState
+    with _ComponentEntryStateMixin<Map<String, String>> {
   @override
   void onElementRegistered(Element element, ComponentEntry entry) {
     if (!entry.isIsland) return;
@@ -180,7 +186,8 @@ class _IslandsDocumentState extends _ManualDocumentState with _ComponentEntrySta
     _islands.add({
       'id': id,
       'name': entry.name,
-      if (entry.params != null) 'params': kDebugMode ? entry.params : stateCodec.encode(entry.params),
+      if (entry.params != null)
+        'params': kDebugMode ? entry.params : stateCodec.encode(entry.params),
     });
   }
 
@@ -193,7 +200,11 @@ class _IslandsDocumentState extends _ManualDocumentState with _ComponentEntrySta
 
   @override
   String? get scriptName {
-    var islands = DocumentBinding.instance!._registryElements.entries.where((e) => e.value.isIsland).toList();
+    var islands = (context.binding as DocumentBinding)
+        ._registryElements
+        .entries
+        .where((e) => e.value.isIsland)
+        .toList();
 
     if (islands.isEmpty) {
       print("[WARNING] Used Document.islands() but no island components were provided.");