+++ packages/jaspr/lib/src/server/server_app.dart
@@ -7,7 +7,7 @@ import 'package:shelf/shelf_io.dart' as shelf_io;
 import '../../server.dart';
 import 'server_handler.dart';
 
-typedef SetupFunction = void Function();
+typedef SetupFunction = void Function(ServerAppBinding binding);
 
 /// An object to be returned from runApp on the server and provide access to the internal http server
 class ServerApp {
@@ -61,7 +61,8 @@ class ServerApp {
     Future.microtask(() async {
       _running = true;
 
-      var handler = createHandler((_, render) => render(_setup), middleware: _middleware, fileHandler: _fileHandler);
+      var handler = createHandler((_, render) => render(_setup),
+          middleware: _middleware, fileHandler: _fileHandler);
 
       if (kDevHotreload) {
         await _reload(this, () {
@@ -108,7 +109,8 @@ Future<void> _reload(ServerApp app, FutureOr<HttpServer> Function() init) async 
     print('[INFO] Server hot reload is enabled.');
   } on StateError catch (e) {
     if (e.message.contains('VM service not available')) {
-      print('[WARNING] Server hot reload not enabled. Run with --enable-vm-service to enable hot reload.');
+      print(
+          '[WARNING] Server hot reload not enabled. Run with --enable-vm-service to enable hot reload.');
     } else {
       rethrow;
     }