+++ packages/jaspr/lib/src/server/server_binding.dart
@@ -3,21 +3,12 @@ import 'dart:convert';
 
 import '../foundation/basic_types.dart';
 import '../foundation/binding.dart';
-import '../foundation/scheduler.dart';
-import '../foundation/sync.dart';
 import '../framework/framework.dart';
 import 'document/document.dart';
 import 'markup_renderer.dart';
 
 /// Global component binding for the server
-class AppBinding extends BindingBase with SchedulerBinding, ComponentsBinding, SyncBinding, DocumentBinding {
-  static AppBinding ensureInitialized() {
-    if (ComponentsBinding.instance == null) {
-      AppBinding();
-    }
-    return ComponentsBinding.instance! as AppBinding;
-  }
-
+class ServerAppBinding extends AppBinding with ComponentsBinding, DocumentBinding {
   @override
   Uri get currentUri => _currentUri!;
   Uri? _currentUri;
@@ -44,7 +35,7 @@ class AppBinding extends BindingBase with SchedulerBinding, ComponentsBinding, S
   Future<String> render() async {
     await rootCompleter.future;
 
-    var renderer = rootElements.values.first.renderer as MarkupDomRenderer;
+    var renderer = rootElement!.renderer as MarkupDomRenderer;
 
     return renderDocument(renderer);
   }
@@ -67,6 +58,7 @@ class AppBinding extends BindingBase with SchedulerBinding, ComponentsBinding, S
 
   @override
   void scheduleFrame(VoidCallback frameCallback) {
-    throw UnsupportedError('Scheduling a frame is not supported on the server, and should never happen.');
+    throw UnsupportedError(
+        'Scheduling a frame is not supported on the server, and should never happen.');
   }
 }