+++ packages/jaspr/lib/src/server/server_renderer.dart
@@ -6,7 +6,8 @@ import 'server_app.dart';
 import 'server_binding.dart';
 
 /// This spawns an isolate for each render, in order to avoid conflicts with static instances and multiple parallel requests
-Future<String> renderHtml(SetupFunction setup, Uri requestUri, Future<String> Function(String) loadFile) async {
+Future<String> renderHtml(
+    SetupFunction setup, Uri requestUri, Future<String> Function(String) loadFile) async {
   var port = ReceivePort();
 
   var message = _RenderMessage(setup, requestUri, port.sendPort);
@@ -50,20 +51,20 @@ class _RenderMessage {
 
 /// Runs the app and returns the rendered html
 void _renderHtml(_RenderMessage message) async {
-  AppBinding.ensureInitialized()
+  var binding = ServerAppBinding()
     ..setCurrentUri(message.requestUri)
     ..setSendPort(message.sendPort);
-  message.setup();
+  message.setup(binding);
 
-  var html = await AppBinding.ensureInitialized().render();
+  var html = await binding.render();
   message.sendPort.send(html);
 }
 
 /// Runs the app and returns the preloaded state data as json
 void _renderData(_RenderMessage message) async {
-  AppBinding.ensureInitialized().setCurrentUri(message.requestUri);
-  message.setup();
+  var binding = ServerAppBinding()..setCurrentUri(message.requestUri);
+  message.setup(binding);
 
-  var data = await AppBinding.ensureInitialized().data();
+  var data = await binding.data();
   message.sendPort.send(data);
 }