+++ packages/jaspr/test/framework/inherited_component/inherited_component_test.dart
@@ -6,19 +6,14 @@ import 'inherited_component_app.dart';
 
 void main() {
   group('inherited component test', () {
-    late ComponentTester tester;
-
-    setUp(() {
-      tester = ComponentTester.setUp();
-    });
-
-    test('should inherit component', () async {
+    testComponents('should inherit component', (tester) async {
       var controller = await tester.pumpTestComponent(App());
 
       // phase 1: inherited component should be mounted
       expect(find.text('Inherited value: 0'), findsOneComponent);
 
-      var state = (find.byType(MyChildComponent).evaluate().first as StatefulElement).state as MyChildState;
+      var state =
+          (find.byType(MyChildComponent).evaluate().first as StatefulElement).state as MyChildState;
 
       // lifecycle: state should be initialized and built a first time
       expect(state.lifecycle, equals(['initState', 'didChangeDependencies', 'build']));
@@ -37,7 +32,9 @@ void main() {
       expect(find.text('Inherited value: 1'), findsOneComponent);
 
       // inherited value should be updated, but without notifying dependants
-      expect(find.byComponentPredicate((component) => component is MyInheritedComponent && component.value == 2),
+      expect(
+          find.byComponentPredicate(
+              (component) => component is MyInheritedComponent && component.value == 2),
           findsOneComponent);
 
       // lifecycle: state should not be updated