+++ packages/jaspr/test/framework/stateful_reparenting/stateful_reparenting_test.dart
@@ -6,20 +6,15 @@ import 'stateful_reparenting_app.dart';
 
 void main() {
   group('stateful reparenting test', () {
-    late ComponentTester tester;
-
-    setUp(() {
-      tester = ComponentTester.setUp();
-    });
-
-    test('should keep state on reparenting', () async {
+    testComponents('should keep state on reparenting', (tester) async {
       var controller = await tester.pumpTestComponent(App());
 
       // phase 1: component should be mounted directly
       expect(find.byType(MyStatefulComponent), findsOneComponent);
       expect(find.tag('div'), findsNothing);
 
-      var state = (find.byType(MyStatefulComponent).evaluate().first as StatefulElement).state as MyState;
+      var state =
+          (find.byType(MyStatefulComponent).evaluate().first as StatefulElement).state as MyState;
 
       // lifecycle: state should be initialized and built a first time
       expect(state.lifecycle, equals(['initState', 'didChangeDependencies', 'build']));
@@ -27,11 +22,14 @@ void main() {
 
       // phase 2: component should be mounted as child of a div element
       await controller.rebuildWith(2);
-      expect(find.descendant(of: find.tag('div'), matching: find.byType(MyStatefulComponent)), findsOneComponent);
+      expect(find.descendant(of: find.tag('div'), matching: find.byType(MyStatefulComponent)),
+          findsOneComponent);
 
       // lifecycle: state should be reparented, updated and built again
       expect(
-          state.lifecycle, equals(['deactivate', 'activate', 'didUpdateComponent', 'didChangeDependencies', 'build']));
+          state.lifecycle,
+          equals(
+              ['deactivate', 'activate', 'didUpdateComponent', 'didChangeDependencies', 'build']));
       state.lifecycle.clear();
 
       // phase 3: component should be unmounted