+++ packages/jaspr_riverpod/test/context_watch_test.dart
@@ -9,13 +9,7 @@ final autoDisposeCounter = StateProvider.autoDispose((ref) => 0);
 
 void main() {
   group('context.watch', () {
-    late ComponentTester tester;
-
-    setUp(() {
-      tester = ComponentTester.setUp();
-    });
-
-    test('context.watch returns provider state and rebuilds on change', () async {
+    testComponents('context.watch returns provider state and rebuilds on change', (tester) async {
       await tester.pumpComponent(providerApp((context) sync* {
         yield Button(
           label: '${context.watch(counter)}',
@@ -33,7 +27,7 @@ void main() {
       expect(find.text('1'), findsOneComponent);
     });
 
-    test('context.watch returns overridden provider state', () async {
+    testComponents('context.watch returns overridden provider state', (tester) async {
       await tester.pumpComponent(providerApp((context) sync* {
         yield Builder(builder: (context) sync* {
           yield Button(
@@ -74,7 +68,7 @@ void main() {
       expect(find.text('b 11'), findsOneComponent);
     });
 
-    test('provider is autodisposed when no longer watched', () async {
+    testComponents('provider is autodisposed when no longer watched', (tester) async {
       await tester.pumpComponent(providerApp((context) sync* {
         var showCounter = true;
         yield StatefulBuilder(